Ramos02

+++ main.py
@@ -271,13 +271,16 @@ class MyApp:
 
         except Exception as e:
             message = f"{e}"
+            try:
+                if 'output_file_path' in locals() and os.path.exists(output_file_path):
+                    os.remove(output_file_path)
+
+            except OSError as e:
+                print(f"Error deleting file: {e}")
             
-            if os.path.exists(output_file_path):
-                os.remove(output_file_path)
+
             
         data_to_return = {'message': message, 'file_path': folder_path, 'status': status}
-
-        sleep(1)
 
         return jsonify(data_to_return)
